Fix Connection lifetime issues with subscriptions
The problem was that even though Subscription holds a reference to the
session, it does not hold a reference to the entire Session class. What
it has is the underlying shared_ptr<sr_sess_ctx_s>. So, after the
Session class gets destroyed, the m_conn field inside the Session class
gets destroyed and everything goes wrong.

There more ways to fix this:
1) Just add an m_conn field eveywhere the m_sess field is. This is the
   easiest way and least clever. You have to remember to put the field
   everywhere, but there aren't going to be many classes I think.
2) Use the deleter for the sr_sess_ctx_s to hold the connection
   shared_ptr. This means that everyone who holds this session will
   automatically prolong the lifetime of the connection. The advantage
   is that you can't forget to have an m_conn field in every class.
   A slight disadvantage is that it's a little bit less readable, than
   having a field in the class. (There is also a slight class size
   advantage, but that's not relevant)
3) Use enable_shared_from_this and save the whole Connection class
   instance in the Session class. This also makes everything
   automatic as in number 2), but makes everything more complex, because
   of shared_ptr.

I chose number 2, because it's the least work, works automatically and
does not require having shared_ptr everywhere.

Change-Id: Ib0eb52fcbe9a017f273586fdf2e3e89cf2dd86fe