Disable SSSD NSS module due to TSAN issues
We're using a regular Linux userland (Fedora 34) in the CI, and that
thing is built without TSAN. That is a problem once a NSS implementation
(which can very well use locking correctly, but in a way that is not
visible to TSAN) gets loaded into the address space. Once that happens,
TSAN flags an error.

Solve this by not using the SSSD package for NSS purposes. This
implementation is more complex than simply reading stuff from files.

Change-Id: I9581303afdd9a45ad9e0a3b4b7c2e031518823b3
Bug: https://github.com/sysrepo/sysrepo/issues/2604