Add Session::getContext
The original C-api attaches this method to the Connection context.
However, the user might not always have the Connection instance
available, for example, in the subscription handlers. One way to solve
this would be to add a getter (Session::getConnection) and implement
Connection::getContext. However, I think that being able to retrieve the
Context from the Session class is very convenient.

Change-Id: Ic5b0a0ff68c7564c91c83e936ebcaf322702e995