Rework processEvents callback
The processEvents callback has been reworked to prevent a specific type
of race and to make working with custom event loops easier. The problem
with the old code was this kind of situation:

The user wants to create multiple custom loop subscriptions through the
same Subscription instance. That means, that the user must save the
Subscription instance somewhere. To create the first subscription, the
user uses the methods in the Session class and saves the return value
to their variable:

auto sub = session.onModuleChange(..., sysrepo::FDHandling{...}, ...);

Now, this is fine at first glance. The user supplies the fd-register
callbacks through the FDHandling struct, the Session::onModuleChange
calls them and then everything is registered just fine. At this point,
the Session::onModuleChange method is almost completed, the FD is
already registered, because the FDHandling callback has been called and
that means that the user eventloop is already listening on the FD. The
eventloop is instructed to call the sub->processEvents() method.

The problem is this: the user could also supply the
SubscribeOptions::Enabled flag. This means that the FD registered by the
register callbacks is immediately available for reading. So, the user
eventloop calls sub->processEvents() as instructed. However, the
Session::onModuleChange might NOT be finished yet at this point, which
means that the `sub` variable might be invalid! This is a data race,
because the eventloop calls a method on the `sub` variable, but that
variable might not be initialized yet.

The user would somehow have to tell the loop to wait, until the variable
is initialized and only then start calling the processEvents method.
This is undesirable, because all custom eventloop users would have to do
that. Instead, the register callback passes the processEvents callback
as a parameter and the user saves it for later use in the eventloop
(when the FD is available for reading).

There is a test. It does reproduce the bug, however, in the new code,
the bug can no longer exist, because the user eventloop does not need to
access the Subscription instance to call processEvents. Also, it was
apparently needed for the `sub` variable to be an std::optional. Now,
I'm not quite sure why is that, but I suppose it assigning to an
optional is less in-place than just having `auto sub = ...`. (it seems
like the `auto` version doesn't call the move callback, but I haven't
investigated.

Side note: there is some extra safety now, because saving the
processEvents callback also prolongs subscription lifetime, because it
("it" being the lambda) captures the subscription shared_ptr.

Change-Id: I3e6945bf788876151a2b9819b254922649413d69
Signed-off-by: Jan Kundrát <jan.kundrat@cesnet.cz>