Fix installation of systemd unit files
The upstream code in sysrpeo obtains the path to systemd unit files via
pkg-config. In order to use them "portably" (e.g., on NixOS, or when
installing into a staging prefix or a fully separate sysroot like in
this CI), one has to pass
`--define-variable=rootdir=${CMAKE_INSTALL_PREFIX}`, but that's not
possible with the current CMake.

Just screw this and pass a relative dir, that way CMake will do the
right thing and combine this path with the `DESTDIR`.

Bug: https://gitlab.kitware.com/cmake/cmake/-/issues/19632
Change-Id: If56c6ea7876745f6eb9a1bdb9a50e4767b624b67