API: prefer std::string over naked C-style strings
This corresponds to the recent changes in libyang-cpp.

!! Beware when porting: in C++20, std::optional<std::string>
unfortunately accepts a std::nullptr which tries to construct a
std::string from a nullptr, which throws std::logic_error. This means
that all calls to, e.g., Session::setItem() have to be audited. This is
"acceptable" so far because we have not committed to any compatibility
with these bindings, yet.

Change-Id: I929e0c60e062a3d06bc2c7b31ceef597427ce64f
Depends-on: https://cesnet-gerrit-czechlight/c/CzechLight/libyang-cpp/+/5565
Depends-on: https://cesnet-gerrit-public/c/CzechLight/libyang-cpp/+/5565
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/libyang-cpp/+/5565