Remove sr_oper_delete_item_str() in favor of sr_delete_item()
Upstream has deprecated the old function, and since the old value
doesn't matter anymore, let's make our API cleaner.

This is an ABI change, hence a version bump. Also, there's been a bunch
of bugs in the original release, so ensure that we're on the most recent
one so that the tests pass.

Change-Id: I231cf90bb198a841ec8bf8fa8f49f604260d5b25
Bug: https://github.com/sysrepo/sysrepo/issues/2952