wrap sr_get_node() for retrieving a single value
Based on popular request, sysrepo got a function for retrieving a single
libyang data node; sometimes it's more efficient compared to
`sr_get_data`.

This function returns a libyang::DataNode and not
std::optional<DataNode> because the C API treats a "node not found" as
an error (and it's documented that way in Doxygen). I think it's a
reasonable decision -- and having a single error path is always nice.

Bug: https://github.com/sysrepo/sysrepo-cpp/issues/13
Change-Id: I9a797c942c638c6a0f7b03c07a1456f8e9b4aba7