tests: measure 10'000 notifications and 10'000 RPCs
This is a very synthetic benchmark; the notifications are
super-lightweight and the RPCs don't do anything, either. The wall time
differs a lot on my T14s gen1, it's anywhere between 3 and 5s for each
part.

The TL;DR version is that the overhead of the RPCs and notifications is
in the sub-ms range on a typical notebook.

Change-Id: Ie2697427d2dc03edce259cc85f592e48814b894f