We don't support sysrepo v3 yet
That version introduced some backward-incompatible changes to the way
how operational data are treated, especially in presence of "discards"
of this data. It's a breaking behavior, and before we release a fix,
let's pin the version to something which still works. The latest sysrepo
semi-release is v 2.12.0, which is unfortunately not tagged.

Change-Id: I3673de9893d6e806926767a72b1ef83ee62bd610