Utilities for working with existing sysrepo:discard-items nodes
findMatchingDiscard() and findMatchingDiscardPrefixes() are useful for
figuring out which nodes need to be prune when adding/modifying a
discarding node.

unlinkFromForest() is useful for removing "a node" from a forest of data
node trees. Sometimes the node to be deleted *could* be a root node
which is also the first sibling among its collection of siblings, and in
that case some extra care is needed to ensure that we still have a
reference to something after we've done with the unlinking.

In the real world this doesn't happen that often because there usually
is something like a top-level container node below which we're
processing these discards. Still, let's put this utility function into a
single place.

Depends-on: https://gerrit.cesnet.cz/c/CzechLight/libyang-cpp/+/8436
Change-Id: Ib346be6887f75220a8a6435864c74aef8d5925bb