add a session getter to dynamic subscriptions
Sometimes it might be useful to get the session which created the
subscription (e.g. to do some work in processEvent callback). There was
no easy way to do it up until now because the session was not stored as
sysrepo::Session.
However, it seems quite easy to change that and add the getter.

Change-Id: Ia943a47debba6f84c3faad5ea48a69da68dac874