fix double call to srsn_terminate
First call of srsn_terminate removes the subscription in sysrepo. A
second call would then throw, because the srsn_terminate would return an
error (the subscription does not exist).

Bear in mind that this does not guard against situations where the
subscription is destroyed by means outside of calling terminate()
method, for instance when stop-time is reached.

Fixes: 415c401 ("wrap dynamic subscription functions")
Change-Id: I5982bc4a93d6277fc6042a7e0604e623482e60a9