throw when a sysrepo::Session is accessed from another thread
Build with -DSYSREPO_CPP_THREAD_CHECKING=OFF to disable this feature.
I'm enabling this by default until profiling shows it to be a
performance hog. I've just wasted significant time of MV's time
debugging a failure which was actually a rather widespread pattern
throughout our codebase -- and that's after using sysrepo for years, and
while being *well* aware of the threading rules. Let's default to safety
here.

The pretty output is only available with GCC 14, and it's only "pretty"
for some definitions of beauty. It prints out decimal numbers which
correspond to the hex thread IDs given in GDB. These are different
numbers from POSIX thread IDs, and also different numbers from the
pretty, short thread sequential numbers shown in GDB. That's life.

Change-Id: I9c3ebeaf92f37f0d77f2cc518c01e2b07141177c