Let Session keep a Connection reference
The Connection class is meant as a lightweight wrapper anyway, so let's
not bother with an explicit shared_ptr<sr_conn_ctx_s>, and keep a
full-blown C++ reference.

The `Session::Session(..., const unmanaged_tag)` is still correct,
AFAICT: it will not attempt to call sr_disconnect() because the
destructor of the shared_ptr<sr_conn_ctx_t> has an empty lambda as its
destructor. That's (still) by design.

Change-Id: I91efc186ff93e7ca27830baa3cbe24ee4c08fba7