Fix a different function return type
The prototype that was specified in the header file did not match what
was used in the corresponding .cpp implementation. This has not been a
problem because that implementation .cpp file was not including the
actual header, so this was not detected. Then, at runtime, this was not
a problem due to the way how derived types are laid out in memory.

Fixes: 6ed5442 Utilities for working with existing sysrepo:discard-items nodes
Change-Id: Ie4cf8c0e39e9c8486a6289c0328510ad6f7360fb